backend/requirements.txt
Django==5.2.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework==3.16.0
django-cors-headers==4.7.0
PyJWT==2.10.1